# cmd/news-alligator/web/docs/swagger.yaml
basePath: /articles
definitions:
  model.Article:
    properties:
      description:
        type: string
      id:
        type: integer
      link:
        type: string
      pubDate:
        type: string
      source:
        $ref: '#/definitions/model.Source'
      title:
        type: string
    type: object
  model.Source:
    properties:
      id:
        type: integer
      link:
        type: string
      name:
        type: string
    type: object
  web.errorResponse:
    properties:
      message:
        type: string
    type: object
host: https://localhost:443
info:
  contact: {}
  description: This is a News Alligator API server.
  title: News Alligator API
  version: "1"
paths:
  /articles:
    get:
      consumes:
      - application/json
      description: Get articles by filter parameters
      operationId: get-articles-by-filter
      parameters:
      - description: Keywords to search for
        in: query
        name: keywords
        type: string
      - description: Sources to search for
        in: query
        name: sources
        type: string
      - description: Start date for search
        in: query
        name: date_start
        type: string
      - description: End date for search
        in: query
        name: date_end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get articles by filter
      tags:
      - articles
  /sources:
    post:
      consumes:
      - application/json
      description: Create a new source
      operationId: create-source
      parameters:
      - description: Source object
        in: body
        name: source
        required: true
        schema:
          $ref: '#/definitions/model.Source'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Source'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create a new source
      tags:
      - sources
  /sources/{id}:
    delete:
      consumes:
      - application/json
      description: Delete source and all associated articles by ID
      operationId: delete-source-by-id
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.errorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete source by ID
      tags:
      - sources
    get:
      consumes:
      - application/json
      description: Immediately fetches news from source by ID
      operationId: fetch-source-by-id
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Fetch source by ID
      tags:
      - sources
    put:
      consumes:
      - application/json
      description: Update source by ID
      operationId: update-source-by-id
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: integer
      - description: Source object
        in: body
        name: source
        required: true
        schema:
          $ref: '#/definitions/model.Source'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Source'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Update source by ID
      tags:
      - sources
swagger: "2.0"