# pkg/model/source.go
package model

import "fmt"

type Source struct {
	Id   int    `json:"id"`
	Name string `json:"name"`
	Link string `json:"link"`
}

func (s Source) String() string {
	return fmt.Sprintf("Source{Id: %d,"+
		" Name: %s,"+
		" Link: %s}", s.Id, s.Name, s.Link)
}